move transaction watcher to network-providers subpackage
