+++ multiversx_sdk/network_providers/__init__.py
@@ -4,5 +4,7 @@ from multiversx_sdk.network_providers.errors import GenericError
 from multiversx_sdk.network_providers.proxy_network_provider import \
     ProxyNetworkProvider
 from multiversx_sdk.network_providers.resources import GenericResponse
+from multiversx_sdk.network_providers.transaction_watcher import \
+    TransactionWatcher
 
-__all__ = ["GenericError", "GenericResponse", "ApiNetworkProvider", "ProxyNetworkProvider"]
+__all__ = ["GenericError", "GenericResponse", "ApiNetworkProvider", "ProxyNetworkProvider", "TransactionWatcher"]
